Add Client32.h and Client64.h
These files should be generated by cgo, but cgo generates files
that are not compatible with cl.exe. For this reason I took the
files generated by cgo, edited them to also work with the win
c compiler and added them as Client32.h and Client64.h. These
are renamed and used in the pipeline.