Do not build for macos-11.0 from pipeline
Decided to not build the client for this macos version as the
GitHub Actions container for it is still in preview and takes a
lot of time to initialize (30 minutes or so). The client for
macos-11.0 will be published manually. The macos-11.0 should
be added back to the pipeline once it is no longer in preview.