Fix publishing and building macos packages
The running delocate on the secrethub macos wheel built
on macos 10.14, 10.15 or 11.0 will always result in a wheel
for macos 10.14, which is by default compatible with the other
two versions. This caused the publish job to fail as it tried
to publish the same wheel multiple times (from the different
matrix entries for the macos versions).

Therefore we can just build the macos wheel on any of the 3
macos versions. I decided to build it on 10.15 as it is not
deprecated and also not in preview.